% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hello_world}
\alias{hello_world}
\title{Say hello!}
\usage{
hello_world(...)
}
\arguments{
\item{...}{Additional things to say, character.}
}
\value{
Logical
}
\description{
Run \code{echo "Hello World!"} in Ubuntu 16.04.
}
